# backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/applications": {
      "get": {
        "operationId": "ApplicationsController_findAll",
        "parameters": [
          {
            "name": "opportunityId",
            "required": false,
            "in": "query",
            "description": "Filter by opportunity ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of applications"
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all applications (recruiter / placement_officer)",
        "tags": [
          "Applications"
        ]
      },
      "post": {
        "operationId": "ApplicationsController_apply",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "required": true,
            "description": "Must be `candidate`",
            "schema": { "type": "string", "enum": ["candidate"] }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Application submitted. If an assessment already exists for the opportunity, candidate also receives a notification."
          },
          "409": {
            "description": "Candidate has already applied to this opportunity"
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Apply to an opportunity (candidate only)",
        "description": "Submits a job/internship application. Use `opportunityId` from `GET /opportunities`. Duplicate applications are rejected with 409.",
        "tags": [
          "Applications"
        ]
      }
    },
    "/applications/my": {
      "get": {
        "operationId": "ApplicationsController_findMine",
        "parameters": [
          {
            "name": "candidateId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Candidate applications"
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get my applications (candidate) — candidateId=1 by default",
        "tags": [
          "Applications"
        ]
      }
    },
    "/applications/stats": {
      "get": {
        "operationId": "ApplicationsController_getStats",
        "parameters": [
          {
            "name": "candidateId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get application statistics for a candidate",
        "tags": [
          "Applications"
        ]
      }
    },
    "/applications/{id}": {
      "get": {
        "operationId": "ApplicationsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get a single application by ID",
        "tags": [
          "Applications"
        ]
      },
      "delete": {
        "operationId": "ApplicationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Delete an application (placement_officer only)",
        "tags": [
          "Applications"
        ]
      }
    },
    "/applications/{id}/status": {
      "patch": {
        "operationId": "ApplicationsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApplicationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update application status (recruiter / placement_officer)",
        "tags": [
          "Applications"
        ]
      }
    },
    "/applications/{id}/withdraw": {
      "patch": {
        "operationId": "ApplicationsController_withdraw",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Withdraw an application (candidate only)",
        "tags": [
          "Applications"
        ]
      }
    },
...
    "/assessments": {
      "get": {
        "operationId": "AssessmentsController_findAll",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "required": true,
            "description": "Use `recruiter` or `placement_officer`",
            "schema": { "type": "string", "enum": ["recruiter", "placement_officer"] }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all scheduled assessments"
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get all assessments (recruiter / placement_officer)",
        "description": "Returns every scheduled assessment. To filter by opportunity, check the `opportunityId` field in the response.",
        "tags": [
          "Assessments"
        ]
      },
      "post": {
        "operationId": "AssessmentsController_create",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "required": true,
            "description": "Must be `recruiter`",
            "schema": { "type": "string", "enum": ["recruiter"] }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssessmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Assessment created and all existing applicants notified automatically"
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Create a new assessment (recruiter only)",
        "description": "Schedules a new assessment linked to an opportunity. **Side-effect:** all candidates who have already applied to the linked `opportunityId` receive an in-app notification automatically.",
        "tags": [
          "Assessments"
        ]
      }
    },
    "/assessments/{id}": {
      "get": {
        "operationId": "AssessmentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Get a single assessment by ID",
        "tags": [
          "Assessments"
        ]
      },
      "put": {
        "operationId": "AssessmentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAssessmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Update an assessment (recruiter only)",
        "tags": [
          "Assessments"
        ]
      },
      "delete": {
        "operationId": "AssessmentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "x-role": []
          }
        ],
        "summary": "Delete an assessment (recruiter only)",
        "tags": [
          "Assessments"
        ]
      }
    },
...
  },
  "info": {
    "title": "CareerNest API",
    "description": "## CareerNest Placement Portal — REST API\n\nFull-stack placement management system built with **NestJS + TypeScript**.\n\n### Authentication\nAll protected endpoints require an `x-role` header. Set it in **Authorize** (top-right) or per-request:\n\n| Role | Who uses it | Example header |\n|---|---|---|\n| `candidate` | Students applying for jobs | `x-role: candidate` |\n| `recruiter` | Company HR posting opportunities | `x-role: recruiter` |\n| `alumni` | Alumni providing referrals | `x-role: alumni` |\n| `placement_officer` | Admin managing the platform | `x-role: placement_officer` |\n\n### Demo Flow\n1. `GET /opportunities` — view live jobs\n2. `POST /applications` — candidate applies\n3. `PATCH /applications/{id}/status` — recruiter shortlists\n4. `POST /assessments` — recruiter schedules test (auto-notifies applicants)\n5. `GET /notifications?candidateId=1` — candidate sees announcement\n\n### Base URL\n`http://localhost:3000`",
    "version": "1.0",
    "contact": {
      "name": "Sai Ganesh & Shammer Basha Shaik",
      "email": "careernest@college.in"
    }
  },
...
  "servers": [
    { "url": "http://localhost:3000", "description": "Local Development Server" }
  ],
  "components": {
    "schemas": {
...
      "CreateApplicationDto": {
        "type": "object",
        "properties": {
          "opportunityId": {
            "type": "number",
            "example": 1,
            "description": "Opportunity ID to apply for"
          },
          "candidateId": {
            "type": "number",
            "example": 1,
            "description": "Candidate user ID"
          },
          "candidateName": {
            "type": "string",
            "example": "Shameer Basha Shaik"
          },
          "branch": {
            "type": "string",
            "example": "Computer Science"
          },
          "cgpa": {
            "type": "number",
            "example": 8.7
          },
          "hasReferral": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "opportunityId"
        ]
      },
      "UpdateApplicationDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "Interview",
            "enum": [
              "Applied",
              "Shortlisted",
              "Assessment",
              "Interview",
              "Offered",
              "Rejected",
              "Withdrawn"
            ]
          },
          "interviewDate": {
            "type": "string",
            "example": "2026-03-15"
          },
          "remarks": {
            "type": "string",
            "example": "Strong technical background"
          }
        }
      },
...
      "CreateAssessmentDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Coding Assessment — Round 1"
          },
          "role": {
            "type": "string",
            "example": "Software Engineer"
          },
          "date": {
            "type": "string",
            "example": "2026-03-05"
          },
          "time": {
            "type": "string",
            "example": "10:00 AM"
          },
          "duration": {
            "type": "string",
            "example": "90 minutes"
          },
          "candidates": {
            "type": "number",
            "example": 12
          },
          "platform": {
            "type": "string",
            "example": "HackerRank"
          },
          "link": {
            "type": "string",
            "example": "https://hackerrank.com/test/x"
          },
          "type": {
            "type": "string",
            "example": "Online Test",
            "enum": [
              "Online Test",
              "Technical Interview",
              "HR Interview",
              "Group Discussion"
            ]
          },
          "isDraft": {
            "type": "boolean",
            "example": false
          },
          "opportunityId": {
            "type": "number",
            "example": 1,
            "description": "Associated opportunity ID"
          }
        },
        "required": [
          "title",
          "role",
          "date",
          "time",
          "duration",
          "candidates",
          "platform",
          "link",
          "type"
        ]
      },
      "UpdateAssessmentDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Coding Assessment — Round 1"
          },
          "role": {
            "type": "string",
            "example": "Software Engineer"
          },
          "date": {
            "type": "string",
            "example": "2026-03-05"
          },
          "time": {
            "type": "string",
            "example": "10:00 AM"
          },
          "duration": {
            "type": "string",
            "example": "90 minutes"
          },
          "candidates": {
            "type": "number",
            "example": 12
          },
          "platform": {
            "type": "string",
            "example": "HackerRank"
          },
          "link": {
            "type": "string",
            "example": "https://hackerrank.com/test/x"
          },
          "type": {
            "type": "string",
            "example": "Online Test",
            "enum": [
              "Online Test",
              "Technical Interview",
              "HR Interview",
              "Group Discussion"
            ]
          },
          "isDraft": {
            "type": "boolean",
            "example": false
          },
          "opportunityId": {
            "type": "number",
            "example": 1,
            "description": "Associated opportunity ID"
          }
        }
      },
...
    },
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "Role of the caller. Valid values: `candidate`, `recruiter`, `alumni`, `placement_officer`"
      }
    }
  }
}